fixed review comments raised by Cristina Ruiz Martin
