+++ .gitignore
@@ -2,3 +2,4 @@ bin/test/*.exe
 bin/top_model/*.exe
 build/test/*.o
 build/top_model/*.o
+build/*.o
