+++ test/data/sender_test_output_transform.txt
@@ -1,58 +1,30 @@
    Time               Value          Port             Component
 ----------------------------------------------------------------------
 00:00:00:000             -1                 out      generator_con
-
 00:00:05:000              0                 out      generator_con
-
 00:00:10:000              0                 out      generator_ack
-
 00:00:15:000              5                 out      generator_con
-
 00:00:25:000              1       packetSentOut            sender1
-
 00:00:25:000             11             dataOut            sender1
-
 00:00:30:000              1                 out      generator_ack
-
 00:00:30:000              1      ackReceivedOut            sender1
-
 00:00:40:000              2       packetSentOut            sender1
-
 00:00:40:000             20             dataOut            sender1
-
 00:01:10:000              2       packetSentOut            sender1
-
 00:01:10:000             20             dataOut            sender1
-
 00:01:30:000              0                 out      generator_ack
-
 00:01:30:000              0      ackReceivedOut            sender1
-
 00:01:40:000              3       packetSentOut            sender1
-
 00:01:40:000             31             dataOut            sender1
-
 00:01:55:000              1                 out      generator_ack
-
 00:01:55:000              1      ackReceivedOut            sender1
-
 00:02:05:000              4       packetSentOut            sender1
-
 00:02:05:000             40             dataOut            sender1
-
 00:02:20:000              1                 out      generator_ack
-
 00:02:35:000              4       packetSentOut            sender1
-
 00:02:35:000             40             dataOut            sender1
-
 00:02:45:000              0                 out      generator_ack
-
 00:02:45:000              0      ackReceivedOut            sender1
-
 00:02:50:000              3                 out      generator_con
-
 00:02:55:000              1                 out      generator_ack
-
 00:02:55:000              1      ackReceivedOut            sender1
-
