+++ test/data/subnet_test_output_transform.txt
@@ -1,80 +1,41 @@
    Time               Value          Port             Component
 ----------------------------------------------------------------------
 00:00:10:000             11                 out          generator
-
 00:00:13:000             11                 out            subnet1
-
 00:00:20:000             20                 out          generator
-
 00:00:23:000             20                 out            subnet1
-
 00:00:30:000             31                 out          generator
-
 00:00:33:000             31                 out            subnet1
-
 00:00:40:000             40                 out          generator
-
 00:00:43:000             40                 out            subnet1
-
 00:00:50:000             51                 out          generator
-
 00:00:53:000             51                 out            subnet1
-
 00:01:00:000             60                 out          generator
-
 00:01:03:000             60                 out            subnet1
-
 00:01:10:000             71                 out          generator
-
 00:01:13:000             71                 out            subnet1
-
 00:01:20:000             80                 out          generator
-
 00:01:23:000             80                 out            subnet1
-
 00:01:30:000             91                 out          generator
-
 00:01:33:000             91                 out            subnet1
-
 00:01:40:000            100                 out          generator
-
 00:01:43:000            100                 out            subnet1
-
 00:01:50:000            111                 out          generator
-
 00:01:53:000            111                 out            subnet1
-
 00:02:00:000            120                 out          generator
-
 00:02:03:000            120                 out            subnet1
-
 00:02:10:000            131                 out          generator
-
 00:02:13:000            131                 out            subnet1
-
 00:02:20:000            140                 out          generator
-
 00:02:23:000            140                 out            subnet1
-
 00:02:30:000            151                 out          generator
-
 00:02:40:000            160                 out          generator
-
 00:02:43:000            160                 out            subnet1
-
 00:02:50:000            171                 out          generator
-
 00:02:53:000            171                 out            subnet1
-
 00:03:00:000            180                 out          generator
-
 00:03:03:000            180                 out            subnet1
-
 00:03:10:000            191                 out          generator
-
 00:03:13:000            191                 out            subnet1
-
 00:03:20:000            200                 out          generator
-
 00:03:23:000            200                 out            subnet1
-
